requirements.txt
tensorflow==2.11.0
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.3
PyYAML==6.0
h5py==3.7.0
python-dotenv==0.21.1